The mod now reads out the Vanilla Crisis level which was set at game start. The Vanilla Crisis level acts as a minimum for the crisis strength. If the mod calculates a higher crisis strength as appropriate, it will add on top of the existing crisis level.
Settings will no longer pop up on game start.

Added a new static offset, which can be used to adjust the calculated crisis level.

Added more fine control when using a custom crisis level.

the mod should now be fine when it's added to a game with an ongoing crisis. The fleet strength is calculated at that moment and crisis buff is applied immediatly.